Created a Separate TileView file for NotesView. Improved the Custom Dialogs by using showGenericDialog. An optional SwitchListTile has been attached in comments to develop it in future for ListView and TileView switching. Some errors were handled.  260825  20:57

+++ lib/views/notes/new_note_view.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:infinity_notes/services/auth/auth_service.dart';
 import 'package:infinity_notes/services/crud/notes_service.dart';
 import 'package:infinity_notes/ui/custom_app_bar.dart';
+import 'package:infinity_notes/ui/dialogs.dart';
 
 class NewNoteView extends StatefulWidget {
   const NewNoteView({super.key});
@@ -21,8 +22,10 @@ class _NewNoteViewState extends State<NewNoteView> {
   late final TextEditingController _titleController;
   late final TextEditingController _textController;
 
-  Timer? _debounce;
+  String _initialTitle = "";
+  String _initialText = "";
 
+  Timer? _debounce;
 
   @override
   void initState() {
@@ -38,26 +41,33 @@ class _NewNoteViewState extends State<NewNoteView> {
     _titleController.addListener(_onTitleChanged);
     _textController.addListener(_onTextChanged);
   }
-  void _onTitleChanged(){
+
+  void _onTitleChanged() {
     setState(() {});
     _handleChange();
   }
-  void _onTextChanged(){
+
+  void _onTextChanged() {
     setState(() {});
     _handleChange();
   }
+
   Future<void> _handleChange() async {
-    if(_debounce?.isActive ?? false) _debounce!.cancel();
+    if (_debounce?.isActive ?? false) _debounce!.cancel();
 
     _debounce = Timer(const Duration(milliseconds: 500), () async {
+
+      final title = _titleController.text.trim();
+      final text = _textController.text.trim();
+      // If there's no change in title or text, don't update the DB
+      if ( _note!=null && title == _initialTitle && text == _initialText) return;
+
+      // These lines will be executed if there's a change in title or text
       final currentUser = AuthService.firebase().currentUser!;
       final email = currentUser.email!;
       final owner = await _notesService.getUser(email: email);
-      final title = _titleController.text.trim();
-      final text = _textController.text.trim();
       // Create note once, after user enters any non empty text
-      if ((title.isNotEmpty || text.isNotEmpty)
-          && _note == null) {
+      if ((title.isNotEmpty || text.isNotEmpty) && _note == null) {
         final newNote = await _notesService.createNote(
           owner: owner,
           title: title,
@@ -69,19 +79,19 @@ class _NewNoteViewState extends State<NewNoteView> {
         return;
       }
       if (_note != null) {
-      if(title.isEmpty && text.isEmpty) {
-        await _notesService.deleteNote(id: _note!.id);
-        setState(() {
-          _note = null;
-        });
-        return;
-      }
+        if (title.isEmpty && text.isEmpty) {
+          await _notesService.deleteNote(id: _note!.id);
+          setState(() {
+            _note = null;
+          });
+          return;
+        }
         final updatedNote = await _notesService.updateNote(
           note: _note!,
           title: title,
           text: text,
         );
-      setState(() => _note = updatedNote);
+        setState(() => _note = updatedNote);
       }
     });
   } // Future<void> _handleChange()
@@ -95,17 +105,18 @@ class _NewNoteViewState extends State<NewNoteView> {
       _note = args;
       _titleController.text = _note!.title;
       _textController.text = _note!.text;
+
+      _initialTitle = _note!.title;
+      _initialText = _note!.text;
+    }
+    else{
+      _initialText = "";
+      _initialTitle = "";
     }
   }
 
-
   @override
   void dispose() {
-    //Delete if note is empty
-    // if((_titleController.text.isEmpty && _textController.text.isEmpty)
-    //     && _note != null){
-    //   _notesService.deleteNote(id: _note!.id);
-    // }
     debugPrint("Note Title Before Disposing: ${_titleController.text}");
     debugPrint("Note Text Before Disposing: ${_textController.text}");
     _titleController.dispose();
@@ -115,64 +126,65 @@ class _NewNoteViewState extends State<NewNoteView> {
     debugPrint("Note Title After Disposing: ${_titleController.text}");
     debugPrint("Note Text After Disposing: ${_textController.text}");
   }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: CustomAppBar(
-        title:_titleController.text.isEmpty
+        title: _titleController.text.isEmpty
             ? "New Note"
-            : _titleController.text
-        ,
+            : _titleController.text,
         themeColor: backgroundColor,
         backgroundColor: Colors.black,
         foregroundColor: foregroundColor,
         actions: [
           PopupMenuButton<String>(
-            onSelected: (value) {
+            onSelected: (value) async {
               if (value == "delete" && _note != null) {
-                _notesService.deleteNote(id: _note!.id);
-                Navigator.of(context).pop();
+                final delete = await showDeleteDialog(context: context);
+                if (delete) {
+                  _notesService.deleteNote(id: _note!.id);
+                  if (!mounted) return;
+                  Navigator.of(context).pop();
+                }
               }
             },
             itemBuilder: (context) => const [
-              PopupMenuItem(
-                value: "delete",
-                child: Text("Delete Note"),
-              ),
+              PopupMenuItem(value: "delete", child: Text("Delete Note")),
             ],
           ),
         ],
       ),
 
-        body: Padding(
-              padding: const EdgeInsets.all(16.0),
-              child: ListView(
-                children: [
-                  TextField(
-                    controller: _titleController,
-                    style: const TextStyle(
-                      fontSize: 20.0,
-                      fontWeight: FontWeight.bold,
-                    ),
-                    decoration: const InputDecoration(
-                      border: OutlineInputBorder(),
-                      hintText: "Title",
-                    ),
-                  ),
-                  const SizedBox(height: 22), // 👈 add some space
-                  TextField(
-                    controller: _textController,
-                    keyboardType: TextInputType.multiline,
-                    maxLines: null,
-                    style: const TextStyle(fontSize: 15.0),
-                    decoration: const InputDecoration(
-                      border: OutlineInputBorder(),
-                      hintText: "Text",
-                    ),
-                  ),
-                ],
+      body: Padding(
+        padding: const EdgeInsets.all(16.0),
+        child: ListView(
+          children: [
+            TextField(
+              controller: _titleController,
+              style: const TextStyle(
+                fontSize: 20.0,
+                fontWeight: FontWeight.bold,
               ),
-            )
+              decoration: const InputDecoration(
+                border: OutlineInputBorder(),
+                hintText: "Title",
+              ),
+            ),
+            const SizedBox(height: 22), // 👈 add some space
+            TextField(
+              controller: _textController,
+              keyboardType: TextInputType.multiline,
+              maxLines: null,
+              style: const TextStyle(fontSize: 15.0),
+              decoration: const InputDecoration(
+                border: OutlineInputBorder(),
+                hintText: "Text",
+              ),
+            ),
+          ],
+        ),
+      ),
     );
   }
 }
